requirements.txt
altgraph==0.17.2
APScheduler==3.6.3
autopep8==1.5.7
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.5
cmdstanpy==0.9.5
colorama==0.4.4
convertdate==2.3.2
cycler==0.10.0
Cython==0.29.24
ephem==4.0.0.2
et-xmlfile==1.1.0
fbprophet==0.7.1
future==0.18.2
hijri-converter==2.2.1
holidays==0.11.2
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.4.0
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
matplotlib==3.3.4
numpy==1.19.5
openpyxl==3.0.8
pandas==1.1.5
pefile==2021.9.3
Pillow==8.3.2
pycodestyle==2.8.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyJWT==2.1.0
PyMeeus==0.5.11
pyparsing==2.4.7
pystan==2.17.1.0
python-dateutil==2.8.2
python-telegram-bot==13.12
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pyupbit==0.2.19
pywin32-ctypes==0.2.0
requests==2.26.0
schedule==1.1.0
setuptools-git==1.2
six==1.16.0
toml==0.10.2
tornado==6.1
tqdm==4.62.2
typing-extensions==3.10.0.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.6
websockets==9.1
zipp==3.6.0
